refactor: utils.core methods into utils to align our package conventions
