+++ paccmann_chemistry/utils/__init__.py
@@ -1,3 +1,4 @@
-from .core import (
-    sequential_data_preparation, collate_fn, kl_weight, get_device, cuda, to_np
+from .utils import (  # noqa
+    sequential_data_preparation, collate_fn, kl_weight, get_device, cuda,
+    to_np, disable_rdkit_logging
 )
